agente_vr/requirements.txt
pandas==2.0.3
numpy==1.24.3
openpyxl==3.1.2
python-dotenv==1.0.0
google-generativeai==0.3.0
xlrd==2.0.1
xlsxwriter==3.1.2
colorlog==6.7.0
tqdm==4.65.0
python-dateutil==2.8.2
